Renamed current conditions sub components
